Maze broken
