+++ Net14/MazeCool/Cells/ClosedDoors.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Net14.Maze.Cells
+namespace MazeCool.Cells
 {
     public class ClosedDoors : BaseCell
     {
