+++ Net14/MazeCool/Cells/Mood.cs
@@ -1,4 +1,4 @@
-﻿namespace Net14.Maze.Cells
+﻿namespace MazeCool.Cells
 {
     public enum Mood
     {
