+++ Net14/MazeCool/Cells/OpenDoors.cs
@@ -2,17 +2,16 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Net14.Maze.Cells
+namespace MazeCool.Cells
 {
-    public class Ground : BaseCell
+    public class OpenDoors: BaseCell
+
     {
-        public override char Symbol => '.';
-
-        public Ground(IMazeLevel mazeLevel) : base(mazeLevel)
+        public OpenDoors(IMazeLevel mazeLevel) : base(mazeLevel)
         {
-
         }
 
+        public override char Symbol => '/';
         public override bool TryToStep(IСharacter hero)
         {
             return true;
