+++ Net14/MazeCool/Direction.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Net14.Maze
+namespace MazeCool
 {
     public enum Direction
     {
