+++ Net14/Net14.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.31129.286
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Net14", "Net14\Net14.csproj", "{7B3094C1-3ABD-4528-A2B4-643073C4408B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Net14.Tests", "Net14.Tests\Net14.Tests.csproj", "{C951C890-8B4B-4A06-A9BA-9EBC1B326436}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Net14.Tests", "Net14.Tests\Net14.Tests.csproj", "{C951C890-8B4B-4A06-A9BA-9EBC1B326436}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MazeCool", "MazeCool\MazeCool.csproj", "{F4D576D1-8D98-45FD-AD66-3051B201583A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{C951C890-8B4B-4A06-A9BA-9EBC1B326436}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C951C890-8B4B-4A06-A9BA-9EBC1B326436}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C951C890-8B4B-4A06-A9BA-9EBC1B326436}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F4D576D1-8D98-45FD-AD66-3051B201583A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F4D576D1-8D98-45FD-AD66-3051B201583A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F4D576D1-8D98-45FD-AD66-3051B201583A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F4D576D1-8D98-45FD-AD66-3051B201583A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
