+++ Net14/Net14/Net14.csproj
@@ -6,4 +6,8 @@
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\MazeCool\MazeCool.csproj" />
+  </ItemGroup>
+
 </Project>
