+++ Net14/Net14/Program.cs
@@ -1,11 +1,4 @@
-﻿using Net14.Maze;
-using Net14.Maze.Cells;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Net14
+﻿namespace Net14
 {
     class Program
     {
